requirements.txt
discord.py[speed]==2.0.1
git+https://github.com/Rapptz/discord-ext-menus@fbb8803779373357e274e1540b368365fd9d8074
lxml
psutil
asyncpg>=0.12.0
typing_extensions>=4,<5
click
parsedatetime
lru_dict
python-dateutil
pygit2